+++ apps/react-namespace/src/createNamespaceContext.tsx
@@ -10,6 +10,9 @@ import { createNamespaceHooks } from './utils/createNamespaceHooks'
 export function createNamespaceContext<StoreType extends NamespaceStore<Record<string | symbol, any>>>({
   globalStore: globalStoreDIP,
   localStore,
+  option = {
+    contextThrowNeed: true,
+  },
 }: StoreOption<StoreType['state'], StoreType>) {
   const Context = createContext<StoreType | undefined>(undefined)
   const globalStore = globalStoreDIP && typeof globalStoreDIP === 'function' ? globalStoreDIP() : globalStoreDIP
@@ -18,7 +21,15 @@ export function createNamespaceContext<StoreType extends NamespaceStore<Record<s
     StoreType['state'],
     StoreType,
     undefined
-  >(() => useContext(Context))
+  >(() => {
+    const context = useContext(Context)
+
+    if (option.contextThrowNeed && isNil(context)) {
+      throw new Error('useNamespaceStoreHooks must be used within a NamespaceContext.Provider')
+    }
+
+    return context
+  })
 
   const Provider: FC<{ overwriteStore?: (() => StoreType) | StoreType; children?: ReactNode }> = ({
     overwriteStore,